Remove legacy evaluation filters and related UI components
