+++ evaluations/src/commonMain/composeResources/values/strings.xml
@@ -73,9 +73,6 @@
 	<string name="error_evaluation_subject_missed">Debes seleccionar una materia</string>
 	<string name="error_evaluation_type_missed">Debes seleccionar un tipo</string>
 	<string name="error_evaluation_max_grade_missed">Debes asignar un peso</string>
-	<string name="label_state_pending">Pendientes</string>
-	<string name="label_state_completed">Completadas</string>
-	<string name="label_state_not_grade">Sin nota</string>
 
 	<string name="dialog_title_add_evaluation_grade">Nota obtenida</string>
 	<string name="dialog_title_add_evaluation_max_grade">Peso (%)</string>
