--- evaluations/src/commonMain/kotlin/com/gdavidpb/tuindice/evaluations/domain/model/EvaluationCourseFilter.kt
@@ -1,26 +1,0 @@
-package com.gdavidpb.tuindice.evaluations.domain.model
-
-import com.gdavidpb.tuindice.base.domain.model.Evaluation
-
-class EvaluationCourseFilter(
-	private val subjectCode: String
-) : EvaluationFilter {
-	override fun getLabel(): String {
-		return subjectCode
-	}
-
-	override fun match(evaluation: Evaluation): Boolean {
-		return subjectCode == evaluation.subjectCode
-	}
-
-	override fun equals(other: Any?): Boolean {
-		if (this === other) return true
-		if (other !is EvaluationCourseFilter) return false
-
-		return subjectCode == other.subjectCode
-	}
-
-	override fun hashCode(): Int {
-		return subjectCode.hashCode()
-	}
-}
