+++ evaluations/src/commonMain/kotlin/com/gdavidpb/tuindice/evaluations/domain/model/GetEvaluations.kt
@@ -10,9 +10,7 @@ sealed interface GetEvaluations {
 	data object NoAttempts : GetEvaluations
 
 	data class Content(
-		val originalEvaluations: List<Evaluation>,
-		val filteredEvaluations: List<Evaluation>,
-		val activeFilters: List<EvaluationFilter>,
+		val evaluations: List<Evaluation>,
 		val hasSyncedEvaluations: Boolean,
 		val displayContext: EvaluationDisplayContext
 	) : GetEvaluations
