--- evaluations/src/commonMain/kotlin/com/gdavidpb/tuindice/evaluations/presentation/model/EvaluationFilterGroupItem.kt
@@ -1,5 +1,0 @@
-package com.gdavidpb.tuindice.evaluations.presentation.model
-
-data class EvaluationFilterGroupItem(
-	val items: List<EvaluationFilterChipItem>
-)
