+++ evaluations/src/commonMain/kotlin/com/gdavidpb/tuindice/evaluations/presentation/viewmodel/EvaluationsViewModel.kt
@@ -1,12 +1,8 @@
 package com.gdavidpb.tuindice.evaluations.presentation.viewmodel
 
-import androidx.lifecycle.viewModelScope
 import com.gdavidpb.tuindice.base.domain.repository.EventPublisher
 import com.gdavidpb.tuindice.base.presentation.Mutation
 import com.gdavidpb.tuindice.base.presentation.viewmodel.BaseViewModel
-import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationFilter
-import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.CheckEvaluationFilterActionProcessor
-import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.ClearEvaluationFiltersActionProcessor
 import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.LoadEvaluationsActionProcessor
 import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.OpenAddEvaluationActionProcessor
 import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.OpenEvaluationActionProcessor
@@ -15,21 +11,12 @@ import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.Refresh
 import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.RemoveEvaluationActionProcessor
 import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.SelectEvaluationsWeekActionProcessor
 import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.SetEvaluationGradeActionProcessor
-import com.gdavidpb.tuindice.evaluations.presentation.action.evaluations.UncheckEvaluationFilterActionProcessor
 import com.gdavidpb.tuindice.evaluations.presentation.contract.Evaluations
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.filterIsInstance
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.flow.stateIn
 
 class EvaluationsViewModel(
 	private val loadEvaluationsActionProcessor: LoadEvaluationsActionProcessor,
 	private val refreshEvaluationsActionProcessor: RefreshEvaluationsActionProcessor,
-	private val checkEvaluationFilterActionProcessor: CheckEvaluationFilterActionProcessor,
-	private val uncheckEvaluationFilterActionProcessor: UncheckEvaluationFilterActionProcessor,
-	private val clearEvaluationFiltersActionProcessor: ClearEvaluationFiltersActionProcessor,
 	private val selectEvaluationsWeekActionProcessor: SelectEvaluationsWeekActionProcessor,
 	private val openAddEvaluationActionProcessor: OpenAddEvaluationActionProcessor,
 	private val pickEvaluationGradeActionProcessor: PickEvaluationGradeActionProcessor,
@@ -41,30 +28,11 @@ class EvaluationsViewModel(
 	name = "evaluations",
 	initialState = Evaluations.State.Idle
 ) {
-	private val activeFilters = state
-		.filterIsInstance<Evaluations.State.Content>()
-		.map { content -> content.activeFilters }
-		.distinctUntilChanged()
-		.stateIn(
-			scope = viewModelScope,
-			started = SharingStarted.WhileSubscribed(5000L),
-			initialValue = emptyList()
-		)
-
 	fun loadEvaluationsAction() =
-		sendAction(Evaluations.Action.LoadEvaluations(activeFilters))
+		sendAction(Evaluations.Action.LoadEvaluations)
 
 	fun refreshEvaluationsAction() =
 		sendAction(Evaluations.Action.RefreshEvaluations)
-
-	fun toggleFilterAction(filter: EvaluationFilter, isChecked: Boolean) =
-		if (isChecked)
-			sendAction(Evaluations.Action.CheckEvaluationFilter(filter))
-		else
-			sendAction(Evaluations.Action.UncheckEvaluationFilter(filter))
-
-	fun clearFiltersAction() =
-		sendAction(Evaluations.Action.ClearEvaluationFilters)
 
 	fun selectWeekAction(weekNumber: Int) =
 		sendAction(Evaluations.Action.SelectWeek(weekNumber))
@@ -94,15 +62,6 @@ class EvaluationsViewModel(
 
 			is Evaluations.Action.RefreshEvaluations ->
 				refreshEvaluationsActionProcessor.process(action, sideEffect)
-
-			is Evaluations.Action.CheckEvaluationFilter ->
-				checkEvaluationFilterActionProcessor.process(action, sideEffect)
-
-			is Evaluations.Action.UncheckEvaluationFilter ->
-				uncheckEvaluationFilterActionProcessor.process(action, sideEffect)
-
-			is Evaluations.Action.ClearEvaluationFilters ->
-				clearEvaluationFiltersActionProcessor.process(action, sideEffect)
 
 			is Evaluations.Action.SelectWeek ->
 				selectEvaluationsWeekActionProcessor.process(action, sideEffect)
