+++ evaluations/src/commonMain/kotlin/com/gdavidpb/tuindice/evaluations/ui/screen/EvaluationsScreen.kt
@@ -10,7 +10,6 @@ import androidx.compose.ui.geometry.Rect
 import com.gdavidpb.tuindice.base.ui.view.SealedCrossfade
 import com.gdavidpb.tuindice.base.ui.view.EmptyStateAnimationView
 import com.gdavidpb.tuindice.base.ui.view.ErrorStateAnimationView
-import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationFilter
 import com.gdavidpb.tuindice.evaluations.presentation.contract.Evaluations
 import com.gdavidpb.tuindice.evaluations.ui.view.EvaluationsContentView
 import com.gdavidpb.tuindice.evaluations.ui.view.EvaluationsEmptyView
@@ -35,8 +34,6 @@ fun EvaluationsScreen(
 	onEvaluationClick: (evaluationId: String, evaluationName: String, subjectCode: String) -> Unit,
 	onEvaluationEdit: (evaluationId: String) -> Unit,
 	onEvaluationDelete: (evaluationId: String) -> Unit,
-	onFilterCheckedChange: (filter: EvaluationFilter, isChecked: Boolean) -> Unit,
-	onClearFiltersClick: () -> Unit,
 	onWeekClick: (Int) -> Unit = {},
 	onRetryClick: () -> Unit,
 	scrollEnabled: Boolean = true,
@@ -62,8 +59,6 @@ fun EvaluationsScreen(
 					EvaluationsContentView(
 						state = targetState,
 						onAddEvaluationClick = onAddEvaluationClick,
-						onClearFiltersClick = onClearFiltersClick,
-						onFilterCheckedChange = onFilterCheckedChange,
 						onWeekClick = onWeekClick,
 						onEvaluationClick = onEvaluationClick,
 						onEvaluationEdit = onEvaluationEdit,
