+++ evaluations/src/commonMain/kotlin/com/gdavidpb/tuindice/evaluations/utils/extension/Computation.kt
@@ -1,16 +1,8 @@
 package com.gdavidpb.tuindice.evaluations.utils.extension
 
-import com.gdavidpb.tuindice.base.domain.model.Evaluation
 import com.gdavidpb.tuindice.base.domain.model.EvaluationScheduleMode
 import com.gdavidpb.tuindice.base.domain.model.EvaluationState
 import com.gdavidpb.tuindice.base.utils.currentTimeMillis
-import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationDateFilter
-import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationFilter
-import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationCourseFilter
-import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationStateFilter
-import com.gdavidpb.tuindice.evaluations.presentation.mapper.EvaluationDateTextMapping
-import com.gdavidpb.tuindice.evaluations.presentation.mapper.getLabel
-import com.gdavidpb.tuindice.evaluations.presentation.mapper.toEvaluationDateGroup
 import kotlin.math.roundToInt
 
 fun Double.toSubjectGrade() = when (roundToInt()) {
@@ -34,48 +26,4 @@ fun computeEvaluationState(
 		grade != null && hasDatePassed -> EvaluationState.COMPLETED
 		else -> EvaluationState.PENDING
 	}
-}
-
-fun List<Evaluation>.computeAvailableFilters(
-	pendingLabel: String,
-	completedLabel: String,
-	noGradeLabel: String,
-	dateTextMapping: EvaluationDateTextMapping
-): List<EvaluationFilter> {
-	val statesFilters = listOf(
-		EvaluationStateFilter(
-			label = pendingLabel
-		) { evaluation -> evaluation.state == EvaluationState.PENDING },
-		EvaluationStateFilter(
-			label = completedLabel
-		) { evaluation -> evaluation.state == EvaluationState.COMPLETED },
-		EvaluationStateFilter(
-			label = noGradeLabel
-		) { evaluation -> evaluation.state == EvaluationState.OVERDUE }
-	)
-
-	val subjectsFilters =
-		map { evaluation -> evaluation.subjectCode }
-			.distinct()
-			.map { subjectCode -> EvaluationCourseFilter(subjectCode) }
-
-	val datesFilters =
-		map { evaluation -> evaluation.toEvaluationDateGroup() }
-			.distinct()
-			.map { group ->
-				EvaluationDateFilter(
-					group = group,
-					label = group.getLabel(dateTextMapping)
-				) { evaluation ->
-					evaluation.toEvaluationDateGroup() == group
-				}
-			}
-
-	val filters = mutableListOf<EvaluationFilter>()
-
-	filters.addAll(statesFilters)
-	filters.addAll(subjectsFilters)
-	filters.addAll(datesFilters)
-
-	return filters
 }
