+++ evaluations/src/commonTest/kotlin/com/gdavidpb/tuindice/evaluations/domain/usecase/EvaluationsUseCaseContractTest.kt
@@ -3,7 +3,6 @@ package com.gdavidpb.tuindice.evaluations.domain.usecase
 import app.cash.turbine.test
 import com.gdavidpb.tuindice.base.domain.model.RecordDataPrerequisiteState
 import com.gdavidpb.tuindice.base.utils.currentTimeMillis
-import com.gdavidpb.tuindice.evaluations.domain.model.EvaluationCourseFilter
 import com.gdavidpb.tuindice.evaluations.domain.model.GetEvaluations
 import com.gdavidpb.tuindice.evaluations.domain.usecase.param.GetEvaluationParams
 import com.gdavidpb.tuindice.evaluations.testing.*
@@ -16,7 +15,7 @@ import kotlin.test.assertTrue
 
 class EvaluationsUseCaseContractTest {
 	@Test
-	fun getEvaluationsUseCase_emitsSortedAndFilteredEvaluations() = runTest {
+	fun getEvaluationsUseCase_emitsSortedEvaluations() = runTest {
 		val currentTime = currentTimeMillis()
 		val futureEvaluation = DEFAULT_PENDING_EVALUATION.copy(
 			date = currentTime + ONE_DAY_MILLIS
@@ -24,7 +23,6 @@ class EvaluationsUseCaseContractTest {
 		val pastEvaluation = DEFAULT_COMPLETED_EVALUATION.copy(
 			date = currentTime - ONE_DAY_MILLIS
 		)
-		val filter = EvaluationCourseFilter(futureEvaluation.subjectCode)
 		val useCase = GetEvaluationsUseCase(
 			evaluationRepository = RecordingEvaluationRepository(
 				evaluationsFlow = flowOf(
@@ -42,14 +40,12 @@ class EvaluationsUseCaseContractTest {
 			reportingRepository = RecordingReportingRepository()
 		)
 
-		useCase.execute(flowOf(listOf(filter))).test {
+		useCase.execute(Unit).test {
 			val result = awaitLoadingThenData(this) as GetEvaluations.Content
 			assertEquals(
 				listOf(pastEvaluation, futureEvaluation),
-				result.originalEvaluations
+				result.evaluations
 			)
-			assertEquals(listOf(futureEvaluation), result.filteredEvaluations)
-			assertEquals(listOf(filter), result.activeFilters)
 			awaitComplete()
 		}
 	}
@@ -67,7 +63,7 @@ class EvaluationsUseCaseContractTest {
 			reportingRepository = reportingRepository
 		)
 
-		useCase.execute(flowOf(emptyList())).test {
+		useCase.execute(Unit).test {
 			assertEquals(GetEvaluations.NoAttempts, awaitLoadingThenData(this))
 			assertTrue(reportingRepository.exceptions.isEmpty())
 			awaitComplete()
@@ -88,7 +84,7 @@ class EvaluationsUseCaseContractTest {
 			reportingRepository = RecordingReportingRepository()
 		)
 
-		useCase.execute(flowOf(emptyList())).test {
+		useCase.execute(Unit).test {
 			assertEquals(GetEvaluations.WaitingForRecordData, awaitLoadingThenData(this))
 			awaitComplete()
 		}
@@ -104,7 +100,7 @@ class EvaluationsUseCaseContractTest {
 			reportingRepository = RecordingReportingRepository()
 		)
 
-		useCase.execute(flowOf(emptyList())).test {
+		useCase.execute(Unit).test {
 			assertEquals(GetEvaluations.RecordDataUnavailable, awaitLoadingThenData(this))
 			awaitComplete()
 		}
