+++ evaluations/src/commonTest/kotlin/com/gdavidpb/tuindice/evaluations/ui/view/EvaluationsContentViewUiTest.kt
@@ -5,7 +5,6 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performClick
 import com.gdavidpb.tuindice.base.ui.BaseUiTags
 import com.gdavidpb.tuindice.evaluations.testing.evaluationsContentState
-import com.gdavidpb.tuindice.evaluations.testing.uiAvailableFilters
 import com.gdavidpb.tuindice.evaluations.ui.EvaluationsUiTags
 import com.gdavidpb.tuindice.testkit.ui.assertNodeHidden
 import com.gdavidpb.tuindice.testkit.ui.assertNodeVisible
@@ -25,8 +24,6 @@ class EvaluationsContentViewUiTest {
 			EvaluationsContentView(
 				state = state,
 				onAddEvaluationClick = { addClicks++ },
-				onClearFiltersClick = {},
-				onFilterCheckedChange = { _, _ -> },
 				onEvaluationClick = { _, _, _ -> },
 				onEvaluationEdit = {},
 				onEvaluationDelete = {}
@@ -36,8 +33,6 @@ class EvaluationsContentViewUiTest {
 		assertNodeVisible(EvaluationsUiTags.EvaluationsContentContainer)
 		assertNodeVisible(EvaluationsUiTags.EvaluationsWeekStrip)
 		assertNodeVisible(EvaluationsUiTags.EvaluationsList)
-		assertNodeHidden(EvaluationsUiTags.EvaluationsFiltersContainer)
-		assertNodeHidden(EvaluationsUiTags.EvaluationsClearFiltersFab)
 		assertNodeVisible(EvaluationsUiTags.EvaluationsAddFab)
 
 		onNodeWithTag(EvaluationsUiTags.EvaluationsAddFab).performClick()
@@ -46,15 +41,13 @@ class EvaluationsContentViewUiTest {
 	}
 
 	@Test
-	fun when_hasEvaluationsAndNoActiveFilters_then_hidesClearFabAndShowsAddFab() = runTuIndiceUiTest {
-		val state = evaluationsContentState(activeFilters = emptyList())
+	fun when_hasEvaluations_then_showsAddFab() = runTuIndiceUiTest {
+		val state = evaluationsContentState()
 
 		setTuIndiceTestContent {
 			EvaluationsContentView(
 				state = state,
 				onAddEvaluationClick = {},
-				onClearFiltersClick = {},
-				onFilterCheckedChange = { _, _ -> },
 				onEvaluationClick = { _, _, _ -> },
 				onEvaluationEdit = {},
 				onEvaluationDelete = {}
@@ -64,23 +57,18 @@ class EvaluationsContentViewUiTest {
 		assertNodeVisible(EvaluationsUiTags.EvaluationsList)
 		assertNodeVisible(EvaluationsUiTags.EvaluationsWeekStrip)
 		assertNodeVisible(EvaluationsUiTags.EvaluationsAddFab)
-		assertNodeHidden(EvaluationsUiTags.EvaluationsFiltersContainer)
-		assertNodeHidden(EvaluationsUiTags.EvaluationsClearFiltersFab)
 	}
 
 	@Test
-	fun when_hasNoEvaluations_then_displaysEmptyMatchAndHidesClearFab() = runTuIndiceUiTest {
+	fun when_hasNoEvaluations_then_displaysEmptyMatch() = runTuIndiceUiTest {
 		val state = evaluationsContentState(
-			originalEvaluations = emptyList(),
-			activeFilters = emptyList()
+			originalEvaluations = emptyList()
 		)
 
 		setTuIndiceTestContent {
 			EvaluationsContentView(
 				state = state,
 				onAddEvaluationClick = {},
-				onClearFiltersClick = {},
-				onFilterCheckedChange = { _, _ -> },
 				onEvaluationClick = { _, _, _ -> },
 				onEvaluationEdit = {},
 				onEvaluationDelete = {}
@@ -91,36 +79,6 @@ class EvaluationsContentViewUiTest {
 		assertNodeVisible(EvaluationsUiTags.EvaluationsWeekStrip)
 		assertNodeVisible(BaseUiTags.EmptyViewContainer)
 		assertNodeVisible(EvaluationsUiTags.EvaluationsAddFab)
-		assertNodeHidden(EvaluationsUiTags.EvaluationsClearFiltersFab)
-		assertNodeHidden(EvaluationsUiTags.EvaluationsFiltersContainer)
 		assertNodeHidden(EvaluationsUiTags.EvaluationsList)
-	}
-
-	@Test
-	fun when_activeFiltersExist_then_filterControlsStayHidden() = runTuIndiceUiTest {
-		val state = evaluationsContentState(
-			activeFilters = listOf(uiAvailableFilters().first())
-		)
-		var addClicks = 0
-
-		setTuIndiceTestContent {
-			EvaluationsContentView(
-				state = state,
-				onAddEvaluationClick = { addClicks++ },
-				onClearFiltersClick = {},
-				onFilterCheckedChange = { _, _ -> },
-				onEvaluationClick = { _, _, _ -> },
-				onEvaluationEdit = {},
-				onEvaluationDelete = {}
-			)
-		}
-
-		assertNodeVisible(EvaluationsUiTags.EvaluationsAddFab)
-		assertNodeHidden(EvaluationsUiTags.EvaluationsFiltersContainer)
-		assertNodeHidden(EvaluationsUiTags.EvaluationsClearFiltersFab)
-
-		onNodeWithTag(EvaluationsUiTags.EvaluationsAddFab).performClick()
-
-		assertEquals(1, addClicks)
 	}
 }
